fix: 修复数据类型不匹配错误，统一使用float32

+++ Train_BILSTM.py
@@ -178,10 +178,10 @@ train_losses_mcae2 = []
 # 中文注释：自定义多输入数据集类（本地定义，非Class_.py中的Dataset）
 class Dataset(Dataset):
     def __init__(self, x, y, z, q):
-        self.x = x.to(torch.double)
-        self.y = y.to(torch.double)
-        self.z = z.to(torch.double)
-        self.q = q.to(torch.double)
+        self.x = x.to(torch.float32)
+        self.y = y.to(torch.float32)
+        self.z = z.to(torch.float32)
+        self.q = q.to(torch.float32)
     def __len__(self):
         return len(self.x)
     def __getitem__(self, idx):
@@ -190,9 +190,9 @@ class Dataset(Dataset):
 # 中文注释：用DataLoader批量加载多通道特征数据
 train_loader_u = DataLoader(Dataset(x_recovered, y_recovered, z_recovered, q_recovered), batch_size=BATCHSIZE, shuffle=False)
 
-# 中文注释：初始化MC-AE模型
-net = CombinedAE(input_size=2, encode2_input_size=3, output_size=110, activation_fn=custom_activation, use_dx_in_forward=True).to(device)
-netx = CombinedAE(input_size=2, encode2_input_size=4, output_size=110, activation_fn=torch.sigmoid, use_dx_in_forward=True).to(device)
+# 中文注释：初始化MC-AE模型（使用float32）
+net = CombinedAE(input_size=2, encode2_input_size=3, output_size=110, activation_fn=custom_activation, use_dx_in_forward=True).to(device).float()
+netx = CombinedAE(input_size=2, encode2_input_size=4, output_size=110, activation_fn=torch.sigmoid, use_dx_in_forward=True).to(device).float()
 
 # 启用数据并行
 if torch.cuda.device_count() > 1:
